Move Content component to its own directory
